+++ src/main/java/racingcar/domain/exception/InvalidCarNameException.java
@@ -1,4 +1,4 @@
-package racingcar.domain.input.exception;
+package racingcar.domain.exception;
 
 public class InvalidCarNameException extends IllegalArgumentException {
