+++ src/main/java/racingcar/domain/exception/InvalidNumberException.java
@@ -1,4 +1,4 @@
-package racingcar.domain.input.exception;
+package racingcar.domain.exception;
 
 public class InvalidNumberException extends NumberFormatException {
